page missing date pruner

+++ src/main/java/uk/co/jassoft/markets/workflow/dates/MissingDatePruner.java
@@ -2,6 +2,7 @@ package uk.co.jassoft.markets.workflow.dates;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -10,6 +11,7 @@ import uk.co.jassoft.markets.repository.MissingDateFormatRepository;
 import uk.co.jassoft.markets.utils.article.ContentGrabber;
 
 import java.util.Date;
+import java.util.List;
 
 /**
  * Created by jonshaw on 22/01/15.
@@ -27,14 +29,27 @@ public class MissingDatePruner {
     @Scheduled(cron = "0 0 * * * ?")
     public void prune()
     {
-        // TODO this needs to be paged
-        for(MissingDateFormat missingDateFormat : missingDateFormatRepository.findAll(new PageRequest(0, 100))) {
-            Date publishedDate = contentGrabber.getPublishedDate(missingDateFormat.getMetatag());
+        int pageCount = 0;
 
-            if(publishedDate != null) {
-                missingDateFormatRepository.delete(missingDateFormat.getId());
+        while (true) {
+            Page<MissingDateFormat> missingDateFormatsPage = missingDateFormatRepository.findAll(new PageRequest(pageCount, 100));
+
+            List<MissingDateFormat> missingDateFormats = missingDateFormatsPage.getContent();
+
+            if(missingDateFormats.isEmpty()) {
+                break;
             }
 
+            missingDateFormats.parallelStream().forEach(missingDateFormat -> {
+                Date publishedDate = contentGrabber.getPublishedDate(missingDateFormat.getMetatag());
+
+                if(publishedDate != null) {
+                    missingDateFormatRepository.delete(missingDateFormat.getId());
+                }
+            });
+
+
+            pageCount++;
         }
     }
 }
